% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viralload.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute likelihood}
\usage{
log_likelihood(
  pars,
  infected,
  observed,
  population,
  tested_population,
  rng,
  n_threads = 1L,
  chunk_size = NULL
)
}
\arguments{
\item{pars}{Named list of parameters. Must have elements \code{a_bar},
\code{a_sigma}, \code{b_bar}, \code{b_sigma}, \code{tmax_bar}, \code{tmax_sigma},
\code{log_vlmax_bar}, \code{log_vlmax_sigma}}

\item{infected}{Vector of number of simulated infections. Must
have length \code{observed$size_full} (this is checked)}

\item{observed}{An \code{observed} object, created by \code{prepare_observed}}

\item{population}{Integer, size of population}

\item{tested_population}{Integer, size of tested population}

\item{rng}{A random number pointer, created by
\link[dust:dust_rng_pointer]{dust::dust_rng_pointer}}

\item{n_threads}{Number of threads to run on, if openmp available.
See \link[dust:dust_openmp_threads]{dust::dust_openmp_threads} for guidance}

\item{chunk_size}{The size of the chunks to run the parallel jobs
in.  The optimal value here will vary depending on your problem
and number of threads. '5' seems like a good guess.}
}
\value{
A single numeric log-likelihood value
}
\description{
Compute likelihood
}
