% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viralload.R
\name{rng_init}
\alias{rng_init}
\title{Create RNG object}
\usage{
rng_init(n_streams, seed = NULL)
}
\arguments{
\item{n_streams}{Number of independent streams required}

\item{seed}{The seed (leave \code{NULL} to use R's seed)}
}
\value{
An opaque pointer to pass through to
\link{log_likelihood}
}
\description{
Create a streaming RNG object
}
